Moved new schema to correct location
